+++ src/model.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash'
 import 'reflect-metadata'
-import {scenario, ScenarioFilter, ScenarioMetadataKey, Never, Always} from './scenario'
+import {scenario, ScenarioFilter, ScenarioMetadataKey, Never, Always, ScenarioName} from './scenario'
 import {ValidateMetadataKey, IValidator, ValidationError} from './validator'
 
 export class Model {
@@ -13,10 +13,10 @@ export class Model {
         return _.kebabCase(this.className)
     }
 
-    static DefaultScenario = 'default'
+    static DefaultScenario = Symbol('default')
 
     @Never()
-    scenario: string = Model.DefaultScenario
+    scenario: ScenarioName = Model.DefaultScenario
 
     @Never()
     scenarioDefaultIncluded: boolean = true
@@ -24,9 +24,10 @@ export class Model {
     constructor() {
     }
 
-    isFieldAvailable(field: string): boolean {
+    isFieldAvailable(field: string, checkScenario=true): boolean {
         if (!Reflect.has(this, field))
             return false
+        if (!checkScenario) return true
         const scenarioFilter = Reflect.getMetadata(ScenarioMetadataKey, this, field) as ScenarioFilter
         return scenarioFilter ? scenarioFilter.check(this.scenario) : this.scenarioDefaultIncluded
     }
@@ -46,24 +47,24 @@ export class Model {
         })
     }
 
-    toDocs(fields?: string[]): Object {
+    toDocs(fields?: string[], force=false, ignoreNil=true): Object {
         if (!fields)
             fields = Object.getOwnPropertyNames(this)
         const docs: any = {}
         fields.forEach(field => {
             const value = Reflect.get(this, field)
-            if (!_.isNil(value) && this.isFieldAvailable(field))
+            if ((!_.isNil(value) || !ignoreNil) && this.isFieldAvailable(field, !force))
                 Reflect.set(docs, field, (value instanceof Model) ? value.toDocs() : value)
         })
         return docs
     }
 
-    validate(fields?: string[], defaultValidator?: IValidator): ValidationError {
+    validate(fields?: string[], defaultValidator?: IValidator, force=false): ValidationError {
         const errors: any = {}
         if (!fields)
             fields = Object.getOwnPropertyNames(this)
         fields.forEach(field => {
-            if (this.isFieldAvailable(field)) {
+            if (this.isFieldAvailable(field, !force)) {
                 const value = Reflect.get(this, field)
                 const validator = defaultValidator || Reflect.getMetadata(ValidateMetadataKey, this, field) as IValidator
                 if (validator) {