+++ CHANGELOG.md
@@ -16,3 +16,9 @@ All notable changes to this project will be documented in this file.
 - CHANGELOG.md created.
 - Publishing to github.
 - MIT LICENSE added.
+
+## [0.1.4] - 2017-08-01
+### Modified
+- Use undefined instead of null for all things.
+- Add force, ignoreNil options to Model.toDoc & Model.validate.
+- Add type ScenarioName to support both string and Symbol scenario.